+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master
 
 * Fix event handling for Turbolinks on Internet Explorer (#43).
+* Fix helper inclusion when `ApplicationController` is eager loaded (#85).
 
 ## 2.0.9